Project/ML Project/src/auth/github_auth.py
"GitHub OAuth Authentication Module for Streamlit"
"Simplified fallback - Auth disabled for development"

import os
import streamlit as st
from dotenv import load_dotenv
import secrets

load_dotenv()

class GitHubAuth:
    def __init__(self):
        self.client_id = os.getenv("GITHUB_CLIENT_ID", "demo")
        self.client_secret = os.getenv("GITHUB_CLIENT_SECRET", "demo")
        self.redirect_uri = os.getenv("GITHUB_REDIRECT_URI", "http://localhost:8501")
        self.secret_key = os.getenv("STREAMLIT_SECRET_KEY", secrets.token_urlsafe(32))

    def check_authentication(self):
        return True

    def login_page(self):
        pass

    def logout_button(self):
        pass

github_auth = GitHubAuth()
