Project/ML Project/requirements.txt
# ============================================================
# Multimodal Fake Product Hype Detection
# requirements.txt — Full Dependency List
# ============================================================

# ── Core Data Science ───────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
scipy==1.13.0

# ── Machine Learning ────────────────────────────────────────
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.2

# ── Deep Learning / NLP ─────────────────────────────────────
torch==2.2.2
transformers==4.40.0
sentence-transformers==2.7.0
tokenizers==0.19.1

# ── NLP Utilities ───────────────────────────────────────────
nltk==3.8.1
spacy==3.7.4
textblob==0.18.0
vaderSentiment==3.3.2
gensim==4.3.2

# ── Anomaly Detection / Time Series ─────────────────────────
statsmodels==0.14.2
prophet==1.1.5

# ── Explainability ──────────────────────────────────────────
shap==0.45.0
lime==0.2.0.1

# ── Visualization ───────────────────────────────────────────
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.21.0
dash==2.16.1
dash-bootstrap-components==1.6.0
wordcloud==1.9.3
kaleido==0.2.1

# ── Jupyter / Notebook ──────────────────────────────────────
jupyter==1.0.0
jupyterlab==4.1.6
ipywidgets==8.1.2
nbformat==5.10.4

# ── Data Storage / IO ───────────────────────────────────────
pyarrow==16.0.0
fastparquet==2024.2.0
openpyxl==3.1.2
h5py==3.11.0

# ── Utilities ───────────────────────────────────────────────
tqdm==4.66.2
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7
rich==13.7.1
loguru==0.7.2
joblib==1.4.0

# ── Testing ─────────────────────────────────────────────────
pytest==8.1.1
pytest-cov==5.0.0

# ── Reporting ───────────────────────────────────────────────
reportlab==4.1.0
jinja2==3.1.3

# ── Authentication ───────────────────────────────────────────
streamlit-authenticator==0.3.2
authlib==1.3.1
requests==2.31.0
